% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi-handling.R
\name{doi_add_url}
\alias{doi_add_url}
\title{add url stub to doi string
'https://doi.org/10.1073/pnas.1106834108'}
\usage{
doi_add_url(x)
}
\arguments{
\item{x}{doi}
}
\description{
add url stub to doi string
'https://doi.org/10.1073/pnas.1106834108'
}

