% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi-handling.R
\name{doi_encode}
\alias{doi_decode}
\alias{doi_encode}
\title{encode doi as if url}
\usage{
doi_encode(x)

doi_decode(x)
}
\arguments{
\item{x}{doi}
}
\description{
this makes it more convenient to use in file names
}

