% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi-handling.R
\name{doi_extract}
\alias{doi_extract}
\title{extract doi from a string}
\usage{
doi_extract(x, regex = "\\\\d+\\\\.\\\\d+\\\\/[^\\\\s]+")
}
\arguments{
\item{x}{chr string that contains doi}

\item{regex}{regex pattern used to extract doi.}
}
\description{
extract doi from a string
}
\details{
Display as eg.: "doi:10.1000/182"
 A DOI takes the form 'prefix/suffix' where prefix is of the form 'registry.registrant'
 and the suffix
 is chosen by the registrant and identifies the specific object.
}

